We love you sir

+++ Project/src/school/management/system/Student.java
@@ -16,7 +16,7 @@ public class Student {
 		
 		
 	}
-   public void setGrade(int grade)
+   public void setGrade(double grade)
    {
 	   this.grade=grade;
    }
